locationDetector/requirements.txt
# Core API framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP client for Overshoot API
httpx==0.26.0
aiohttp==3.9.1

# Data validation and serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Image processing
Pillow==10.2.0
python-magic==0.4.27

# Environment management
python-dotenv==1.0.0

# Logging and monitoring
structlog==24.1.0

# Testing (optional but recommended)
pytest==7.4.4
pytest-asyncio==0.23.3
